+++ tests/testthat/test_ne_file_name.R
@@ -1,31 +1,36 @@
 library(httr)
 
+url_exists <- function(url) {
+  url <- sanitize_gdal_url(url)
+  res <- HEAD(url)
+  status_code(res) == 200L
+}
+
 # a non exhaustive list of ne download urls
 # initially taken from those in data_download_script.r
 # fmt: skip
 urls <- c(
-  ne_file_name(scale = 110L, type = "countries", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 110L, type = "map_units", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 110L, type = "sovereignty", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 50L, type = "countries", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 50L, type = "map_units", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 50L, type = "sovereignty", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 10L, type = "countries", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 10L, type = "map_units", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 10L, type = "sovereignty", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 50L, type = "states", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 10L, type = "states", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 110L, type = "tiny_countries", category = "cultural", full_url = TRUE),
-  ne_file_name(scale = 110L, type = "coastline", category = "physical", full_url = TRUE),
-  ne_file_name(scale = 50L, type = "coastline", category = "physical", full_url = TRUE),
-  ne_file_name(scale = 10L, type = "coastline", category = "physical", full_url = TRUE)
+  ne_file_name(scale = 110L, type = "countries", category = "cultural"),
+  ne_file_name(scale = 110L, type = "map_units", category = "cultural"),
+  ne_file_name(scale = 110L, type = "sovereignty", category = "cultural"),
+  ne_file_name(scale = 50L, type = "countries", category = "cultural"),
+  ne_file_name(scale = 50L, type = "map_units", category = "cultural"),
+  ne_file_name(scale = 50L, type = "sovereignty", category = "cultural"),
+  ne_file_name(scale = 10L, type = "countries", category = "cultural"),
+  ne_file_name(scale = 10L, type = "map_units", category = "cultural"),
+  ne_file_name(scale = 10L, type = "sovereignty", category = "cultural"),
+  ne_file_name(scale = 50L, type = "states", category = "cultural"),
+  ne_file_name(scale = 10L, type = "states", category = "cultural"),
+  ne_file_name(scale = 110L, type = "tiny_countries", category = "cultural"),
+  ne_file_name(scale = 110L, type = "coastline", category = "physical"),
+  ne_file_name(scale = 50L, type = "coastline", category = "physical"),
+  ne_file_name(scale = 10L, type = "coastline", category = "physical")
 )
 
 url_expect_fail <- ne_file_name(
   scale = 110L,
   type = "expect_fail",
-  category = "cultural",
-  full_url = TRUE
+  category = "cultural"
 )
 
 test_that("urls for downloads created by the package exist", {
@@ -33,9 +38,7 @@ test_that("urls for downloads created by the package exist", {
   skip_if_not_installed("rnaturalearthdata")
   skip_if_not_installed("rnaturalearthhires")
 
-  # tests all of the urls put into the vector above
-  # info=x means that test fail messages include the failed urls
-  sapply(urls, function(x) expect_false(httr::http_error(x), info = x))
+  expect_true(all(vapply(urls, url_exists, logical(1L))))
 })
 
 test_that("a bogus url does not exist", {
@@ -43,5 +46,5 @@ test_that("a bogus url does not exist", {
   skip_if_not_installed("rnaturalearthdata")
   skip_if_not_installed("rnaturalearthhires")
 
-  expect_true(httr::http_error(url_expect_fail))
+  expect_false(url_exists(url_expect_fail))
 })